Fix 2 money bugs. 1. Not being able to buy when bank exceeds INT_MAX/100. 2. Dropping over INT_MAX gold sets the resulting pile cost to gold * 100 (overflow). This commit fixes the fact that saccing the resulting pile results in a negative silver balance. The negative cost bug is still there. rooflez.com/mantis/view.php?id=275
http://rooflez.com/mantis/view.php?id=274

+++ src/handler.c
@@ -3563,9 +3563,9 @@ void add_money( CHAR_DATA *ch, int gold, int silver, CHAR_DATA *source )
 }
 
 /* deduct cost from a character */
-void deduct_cost(CHAR_DATA *ch, int cost)
+void deduct_cost(CHAR_DATA *ch, long cost)
 {
-    int silver = 0, gold = 0, bank = 0;
+    long silver = 0, gold = 0, bank = 0;
     char buf[MSL];
     
     silver = UMIN(ch->silver,cost);
